src/requirements.txt
black==24.2.0
commitizen==3.18.3
dj-database-url==2.1.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
pytest-django==4.8.0
python-decouple==3.8
whitenoise==6.6.0
